book_recommendations/requirements.txt
numpy==1.21.0
pandas==1.3.3
Flask==2.1.2
scikit-learn==0.24.2
transformers==4.9.2
torch==1.9.0
faiss-cpu==1.7.1
scikit-surprise==1.1.1
textblob==0.15.3
requests==2.26.0
python-dotenv==0.19.0
